Add files via upload
This Python code is a number guessing game where the user tries to guess a randomly generated number within a specified range. The user enters their guesses, and the program provides feedback on whether the guess is above or below the target number. The game continues until the user correctly guesses the number. The code keeps track of the number of guesses made and displays it at the end. Have fun trying to find the hidden number! 🤔💭